Initial implementation of hexagonal archotecture
