+++ src/main/java/com/javateam/medicalcenter/ApplicationStartup.java
@@ -1,0 +1,12 @@
+package com.javateam.medicalcenter;
+
+import org.springframework.boot.CommandLineRunner;
+import org.springframework.stereotype.Component;
+
+@Component
+public class ApplicationStartup implements CommandLineRunner {
+    @Override
+    public void run(String... args) {
+
+    }
+}
