+++ src/main/java/com/javateam/medicalcenter/MedicalCenterApplication.java
@@ -1,6 +1,6 @@
 package com.javateam.medicalcenter;
 
-import com.javateam.medicalcenter.entity.Employee;
+import com.javateam.medicalcenter.employees.domain.Employee;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.support.ClassPathXmlApplicationContext;
