+++ src/main/java/com/javateam/medicalcenter/MedicalCenterConfiguration.java
@@ -1,7 +1,6 @@
 package com.javateam.medicalcenter;
 
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
@@ -11,4 +10,3 @@ public class MedicalCenterConfiguration {
 
   }
 
-}
