--- src/main/java/com/javateam/medicalcenter/account/application/service/GetAccountBalanceService.java
@@ -1,21 +1,0 @@
-package com.javateam.medicalcenter.account.application.service;
-
-import java.time.LocalDateTime;
-
-import io.reflectoring.buckpal.account.application.port.in.GetAccountBalanceQuery;
-import io.reflectoring.buckpal.account.application.port.out.LoadAccountPort;
-import io.reflectoring.buckpal.account.domain.Account.AccountId;
-import io.reflectoring.buckpal.account.domain.Money;
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
-class GetAccountBalanceService implements GetAccountBalanceQuery {
-
-	private final LoadAccountPort loadAccountPort;
-
-	@Override
-	public Money getAccountBalance(AccountId accountId) {
-		return loadAccountPort.loadAccount(accountId, LocalDateTime.now())
-				.calculateBalance();
-	}
-}
