--- src/main/java/com/javateam/medicalcenter/account/application/service/ThresholdExceededException.java
@@ -1,11 +1,0 @@
-package com.javateam.medicalcenter.account.application.service;
-
-import io.reflectoring.buckpal.account.domain.Money;
-
-public class ThresholdExceededException extends RuntimeException {
-
-  public ThresholdExceededException(Money threshold, Money actual) {
-    super(String.format("Maximum threshold for transferring money exceeded: tried to transfer %s but threshold is %s!", actual, threshold));
-  }
-
-}
