+++ src/main/java/com/javateam/medicalcenter/appointments/domain/AppointmentRepository.java
@@ -1,0 +1,14 @@
+package com.javateam.medicalcenter.appointments.domain;
+
+import com.javateam.medicalcenter.appointments.domain.Appointment;
+import com.javateam.medicalcenter.employees.domain.Doctor;
+import com.javateam.medicalcenter.patients.domain.Patient;
+import org.springframework.data.repository.CrudRepository;
+
+public interface AppointmentRepository extends CrudRepository<Appointment, Long> {
+    Iterable<Appointment> findByPatient(Patient patient);
+
+    Iterable<Appointment> findByDoctor(Doctor doctor);
+}
+
+
