--- src/main/java/com/javateam/medicalcenter/common/SelfValidating.java
@@ -1,29 +1,0 @@
-package com.javateam.medicalcenter.common;
-
-import javax.validation.ConstraintViolation;
-import javax.validation.ConstraintViolationException;
-import javax.validation.Validation;
-import javax.validation.Validator;
-import javax.validation.ValidatorFactory;
-import java.util.Set;
-
-public abstract class SelfValidating<T> {
-
-  private Validator validator;
-
-  public SelfValidating() {
-    ValidatorFactory factory = Validation.buildDefaultValidatorFactory();
-    validator = factory.getValidator();
-  }
-
-  /**
-   * Evaluates all Bean Validations on the attributes of this
-   * instance.
-   */
-  protected void validateSelf() {
-    Set<ConstraintViolation<T>> violations = validator.validate((T) this);
-    if (!violations.isEmpty()) {
-      throw new ConstraintViolationException(violations);
-    }
-  }
-}
