--- src/main/java/com/javateam/medicalcenter/entity/Patient.java
@@ -1,16 +1,0 @@
-package com.javateam.medicalcenter.entity;
-
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.Entity;
-import javax.persistence.OneToMany;
-import java.util.List;
-
-@Entity
-@Data
-@NoArgsConstructor
-public class Patient extends User {
-    @OneToMany(mappedBy = "patient")
-    private List<Appointment> appointments;
-}
