+++ src/main/java/com/javateam/medicalcenter/users/domain/User.java
@@ -1,0 +1,29 @@
+package com.javateam.medicalcenter.users.domain;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.*;
+import java.sql.Time;
+import java.util.List;
+
+@Entity
+@Table(name="USERS")
+@Inheritance(strategy= InheritanceType.JOINED)
+@DiscriminatorColumn(name = "USER_TYPE", discriminatorType = DiscriminatorType.STRING)
+// W dziedziczących klasach dodajemy @DiscriminatorValue("nazwa_typu_usera") nad każdą klasą
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public abstract class User {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "ID")
+    private Long id;
+    @Column(name = "LOGIN")
+    private String login;
+    @Column(name = "PASSWORD")
+    private String password;
+
+}
